Move time words in its own file
